Allow float comparison between nic names in isys.py. (#246135)

+++ isys/isys.py
@@ -697,8 +697,8 @@ def compareDrives(first, second):
 
 def compareNetDevices(first, second):
     try:
-        trimmed_first = int(first.lstrip(string.letters))
-        trimmed_second = int(second.lstrip(string.letters))
+        trimmed_first = float(first.lstrip(string.letters))
+        trimmed_second = float(second.lstrip(string.letters))
     except:
         return 0
